Revert "Made new test file / broke out tests"
This reverts commit 19619239e50143458560a04a2abb1f921c4308e7.
